% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_bin_counts.R
\name{aggregate_bin_counts}
\alias{aggregate_bin_counts}
\title{Aggregate bin summary counts into a matrix of counts format}
\usage{
aggregate_bin_counts(dir, breaks = NULL, flanking_bases = 2)
}
\arguments{
\item{dir}{The directory containing the files that are to be read and aggregated over.}

\item{breaks}{The breaks used for the \code{damage.build.counts} function.}

\item{flanking_bases}{The number of flanking bases. Defaults to 2.}
}
\value{
Returns a matrix of aggregate counts of each signature for each file arranged along a
row of the matrix. The number of rows correspond to the number of files in the directory
and the number of columns equalling the number of distinct mutation signatures.
}
\description{
For each file of the directory, performs the \code{damage.build.counts}
function and aggregates the data into counts matrix format with the number of rows
equal to the number of files in the directory and the columns corresponding to the
number of different mutation signatures recorded. The files must be in .csv format.
}
\keyword{aggregate_counts}

