% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/club_signature_counts_2.R
\name{club_signature_counts_2}
\alias{club_signature_counts_2}
\title{Club signatures accounting for the strand break and strand effect}
\usage{
club_signature_counts_2(signature_counts, flanking_bases = 1)
}
\arguments{
\item{signature_counts}{The matrix of counts of all signatures as produced by \code{aggregate_bin_counts}.}

\item{flanking_bases}{The number of flanking bases. Defaults to 2.}
}
\value{
Returns a matrix of clubbed signatures. The default choice of the mutation
signatures are C->T, C->A, C->G, T->A, T->C and T->G. The other sets of mutations
like G->A are clubbed to C->T to remove strand bias.
}
\description{
Function for clubbing signatures from the two strands to remove
strnad bias. An example is G->A and C->T are clubbed into C->T to remove the bias
due to whether 5' or 3' ends of the strands has been sequenced.
}
\keyword{aggregate_counts}

