% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/club_signature_validation_plot.R
\name{club_signature_validation_plot}
\alias{club_signature_validation_plot}
\title{Function to check if the signatures should be clubbed}
\usage{
club_signature_validation_plot(signature_counts, flanking_bases = 2,
  cex = 1, col = "red", pch = 20, xlab = "C->T", ylab = "G->A",
  lty = 1, lwd = 1, mar = c(5, 4, 4, 4), log = FALSE)
}
\arguments{
\item{signature_counts}{The matrix of counts of all signatures as produced by \code{aggregate_bin_counts}.}

\item{flanking_bases}{The number of flanking bases. Defaults to 2.}

\item{cex}{The size of the dots used for plotting}

\item{col}{The color of the dot}

\item{pch}{The shape of the dot}

\item{xlab}{The X-axis label}

\item{ylab}{The Y-axis label}

\item{lty}{The line type of the dot}

\item{lwd}{The line width of the dot}

\item{mar}{The margin of the plot}

\item{log}{if TRUE, the lof of the counts are plotted, else plot on original scale. Default is FALSE.}
}
\value{
Returns a plot of number of C->T versus number of G->A mutations for each sample.
}
\description{
Validity check whether the C->T and G->A signatures should be clubbed. This function
reads in the matrix of signature counts produced by the \code{aggregate_bin_counts} and plots the
C->T and G->A mutations for each sample.
}
\keyword{plot_signatures}

