% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damage.build.counts.R
\name{damage_build_bin_counts}
\alias{damage_build_bin_counts}
\title{Summarize the mutationFileFormat data into bins along the read length}
\usage{
damage_build_bin_counts(file, breaks = NULL, type = 1)
}
\arguments{
\item{file}{a MutationFile format data obtained from applying
\code{generate_summary_data} on the BAM files.}

\item{breaks}{The breaks along the read length used to form the bins. The default os
is NULL in which case, we choose the default option of having evenly separated
bins defined at spacings of 5 along the read length.}

\item{type}{Two designs of output matrices available, according to user preference.}
}
\value{
Returns a matrix with columns representing the signature and the bins
in which the signature belongs to and a column representing the counts of the
occurrences of that signature in that bin  across the entire BAM file.
}
\description{
Build a summary counts table by defining bins along the read length
based on the distance from the ends of the read and then sumaarizing the mutationFile
format data into bins by counting the number of occurrences of reads into each
bin.
}
\keyword{summarize_counts}

