% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/damageLogo.R
\name{damageLogo}
\alias{damageLogo}
\title{Builds damage logo plots for the clusters}
\usage{
damageLogo(theta, sig_names = NULL, ic.scale = TRUE, yscale_change = TRUE,
  xaxis = TRUE, yaxis = TRUE, xaxis_fontsize = 10, xlab_fontsize = 15,
  y_fontsize = 15, mut_width = 2, start = 1e-04,
  pop_names = paste0("Cluster ", 1:dim(theta)[2]))
}
\arguments{
\item{theta}{The theta matrix obtained from running the grade of membership model that stores for each cluster, the
probability distribution over all the mutational signatures.}

\item{sig_names}{The mutational signature names. Defaults to the rownames of the theta matrix above.}

\item{ic.scale}{A binary variable indicating whether the height of the bars for substitution and flanking bases should be
adjusted by the information criterion.}

\item{yscale_change}{A binary variable indicating whether the Y axis scale should be adjusted based on the size of the
logos, defaults to TRUE.}

\item{xaxis}{A binary indicating whether the X axis of the logo plot should be shown}

\item{yaxis}{A binary indicating whether the Y axis of the logo plot should be shown}

\item{xaxis_fontsize}{The fontsize of the X axis ticks.}

\item{xlab_fontsize}{The fontsize of the X axis labels.}

\item{y_fontsize}{The fontsize of the Y axis ticks.}

\item{mut_width}{Thw width of the bar for the mutation at the center.}

\item{start}{The starting point of the stacking of logos on the Y axis. Should be close to 0, defau;ts to 0.0001.}

\item{pop_names}{The title of the plot. Defaults to the cluster labels.}
}
\value{
Returns logo plots for each cluster
}
\description{
Damage Logo plots for each cluster representing the substitution frequency of the 6 substitutional
patterns (adjusting for strand bias) and the flanking bases arranged as per relative frequency on either side of
the substitution.
}

