% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_signatures_by_location.R
\name{filter_signatures_by_location}
\alias{filter_signatures_by_location}
\title{Function to extract signature counts of mutations occurring only upto a certain position along the genome}
\usage{
filter_signatures_by_location(mat, max_pos = 23, flanking_bases = 2)
}
\arguments{
\item{mat}{The matrix of the signature counts obtained by \code{aggregate_bin_counts} or
\code{club_signature_counts} functions.}

\item{max_pos}{The highest position along the read from the start for which the counts are extracted.}

\item{flanking_bases}{The number of flanking bases in the signatures}
}
\value{
Returns a filtered matrix of signature counts of mutations upto a certain distance from ends of the read.
}
\description{
This function can be used by the user to extract the counts data for mutations occurring
from the end of the read upto a certain maximum distance.
}
\keyword{filter-signatures}

