% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_signatures_per_substitution.R
\name{filter_signatures_per_substitution}
\alias{filter_signatures_per_substitution}
\title{Function to extract signature counts only for a specific substitution pattern}
\usage{
filter_signatures_per_substitution(counts, pattern, flanking_bases = 2,
  use_prop = FALSE)
}
\arguments{
\item{counts}{The matrix of the signature counts obtained by \code{aggregate_bin_counts} or
\code{club_signature_counts} functions.}

\item{pattern}{The substitution pattern for which the user intends to extract the counts.}

\item{flanking_bases}{The number of flanking bases for the mutation signatures.}

\item{use_prop}{If TRUE, we record the proportion of the signatures with that pattern
compared to signatures with all patterns for each sample. Defaults to FALSE.}
}
\value{
Returns a filtered matrix of signature counts for a specific substitution pattern.
}
\description{
This function can be used by the user to extract the counts data only for mutation signatures with
a specific substitution pattern (for example C -> T).
}
\keyword{filter-signatures}

