% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_signatures_w_mutation.R
\name{filter_signatures_w_mutation}
\alias{filter_signatures_w_mutation}
\title{Function to extract counts for only mutation with no flanking bases.}
\usage{
filter_signatures_w_mutation(counts, flanking_bases = 2)
}
\arguments{
\item{counts}{The matrix of the signature counts obtained by \code{aggregate_bin_counts} or
\code{club_signature_counts} functions.}

\item{flanking_bases}{The number of flanking bases for the mutation signatures}
}
\value{
Returns a filtered matrix with same number of rows (samples) but reduced set of columns that only contains
the counts for each mutation without the flanking bases and location information.
}
\description{
In case the user wants to disregard the position information and flanking bases of the mutational
signatures, this function allows the user to filter the data by keeping only the mutation information and
aggregating the counts for each mutation across all the flanking bases and the positional information.
}
\keyword{filter-signatures,}
\keyword{mutation}

