% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridPCA_signatures.R
\name{gridPCA_signatures}
\alias{gridPCA_signatures}
\title{Function to perform grid plots of PCA for signature counts}
\usage{
gridPCA_signatures(counts, labs)
}
\arguments{
\item{labs}{The factor labels for the samples used for coloring in the PC plot}

\item{signature_counts}{The matrix of the signature counts}
}
\value{
Returns grid plot of the PCA plots
}
\description{
This fucntion performs PCA on normalized signature counts data
obtained from \code{aggregate_bin_counts} or \code{club_signature_counts}
and plots the scatter plots for PC1 vs PC2, PC2 vs PC3 and PC1 vs PC3 in a grid plot.
}
\keyword{PCA_signatures}

