% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_plot_damage_full.R
\name{pattern_plot_full}
\alias{pattern_plot_full}
\title{Plot mutation signature patterns across read length}
\usage{
pattern_plot_full(file, pattern, plot_type = c("left", "right", "both"),
  sample_name = NULL, cols = c("black", "red", "green", "blue", "orange",
  "purple"), legend_cex = 0.5)
}
\arguments{
\item{file}{a MutationFile format data obtained from applying
\code{generate_summary_data} on the BAM files.}

\item{pattern}{the mutation signature pattern, whose frequency along the
read, the user wants to plot.}

\item{plot_type}{Three options are - "left", "right" and "both" demonstrating
whether the user wants the frequency plot of the pattern at the left end of the
read, right end of the read or both ends.}

\item{sample_name}{The name of the BAM file to be used in title. Defaults to NULL.}

\item{cols}{The color vector of the mutational patterns plotted. The number of colors must be more than the number of
mutations used.}

\item{legend_cex}{The size of the legend which highlight the names of the patterns}
}
\value{
Returns a frequency plot of the pattern across the read length.
}
\description{
This function reads in the MutationFile format data comprising
of the type of  mutation signatures and its location information from the
ends of reads, and plots a line plot of the frequency of a pattern provided
by the user, along the read length.
}
\keyword{mutation}
\keyword{pattern,}
\keyword{signature}

