% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strand_breaks_composition.R
\name{strand_breaks_composition}
\alias{strand_breaks_composition}
\title{Computes strand break composition for mutation signatures at the extremes}
\usage{
strand_breaks_composition(file, flanking_bases = 2)
}
\arguments{
\item{file}{The file name for which the strand break composition is sought}

\item{flanking_bases}{The number of flanking bases in the signatures}
}
\value{
Returns a list with two elements, each being a vector of counts of nucleotides mapped to the left and right strand breaks
}
\description{
This function computes the composition of nucleotides at the strand breaks positions corresponding
to mutation signatures involving strand breaks.
}
\keyword{filter-signatures}

